+++ app/models/page_report.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# Tracks position of a report on a page.
+class PageReport < ApplicationRecord
+  # Relationships
+  belongs_to :page
+  belongs_to :report
+end