+++ test/models/page_report_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class PageReportTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end